Redo the routing and data system
After doing some thinking and asking some people about their opinions, I realized that the old system had some issues, such as:
- being unable to easily attach middleware to only one specific method,
- no way to run middleware only if a node actually matches, as express users expected,
- middleware wasn't being sorted, resulting in middleware from different sources having to run one after another regardless of stages,
- lack of distinction between middleware and terminating middleware.

All those things (and probably some other stuff I already forgot about) resulted in a loss of my will to live and a major rewrite of a lot of the code which took a bunch of time and hurt my brain. Fortunately some logic carried over, but there were still a lot of changes.

The current system should be very flexible (though paramaters are not yet implemented), and should feel familiar to express developers, or at least I'm hoping it will. I haven't really done any field studies nor tried using it myself much yet, but I thought about the design a bunch and it should work!

So yeah... For now I'm committing this because it finally passess all the tests (after several days of failing a bunch of the Router ones) and I want that success to be saved and not disappear by accident.