Fix potential vulnerability allowing middleware to be ran twice
Interally the strings `middleware` and `all` were used as keys in the same Map as the one storing middleware to be ran on specific HTTP request methods. According to the HTTP RFC, uppercase method names are only a recommendation, meaning that while Node.js was rejecting requests using those as the method with `400 Bad Request`, there's no reason they couldn't eventually become valid or have been valid in the past. By prepending a null byte before the `SpecialMethod` values, we prevent the parser from ever giving us a request with a dangerous value, fixing this issue.