The name shall be koa-butterfly
The name was inspired by an actual butterfly that flew into my room on 2020-09-16 while I was working on this router. It seems like a nice name, too.

If you're reading this and wondering about the butterfly, it had reddish-brown wings with yellow and light pink circles, and liked the underside of my keyboard shelf. After flying in at some point past 17, it settled and stayed there for at least three days. It stayed in my room until 2020-09-21 09:53, when I let it out after it woke me up with its flapping wings hitting the window blinds. It's a nice memory that stayed in my mind for long enough that I decided to name the library after the butterfly.